refactor(generate): now just a function
- also updates tests
